Factor out data providers common across functional and integration tests (#590)
