+++ tests/DataProvider/CreateSourceInvalidRequestDataProviderTrait.php
@@ -1,0 +1,53 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\DataProvider;
+
+use App\Enum\Source\Type;
+use App\Request\InvalidSourceTypeRequest;
+use App\Request\SourceRequestInterface;
+
+trait CreateSourceInvalidRequestDataProviderTrait
+{
+    /**
+     * @return array<mixed>
+     */
+    public function createSourceInvalidRequestDataProvider(): array
+    {
+        return [
+            'invalid source type' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => 'invalid',
+                ],
+                'expectedResponseData' => [
+                    'error' => [
+                        'type' => 'invalid_request',
+                        'payload' => [
+                            'type' => [
+                                'value' => 'invalid',
+                                'message' => InvalidSourceTypeRequest::ERROR_MESSAGE,
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+            'git source missing host url' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
+                ],
+                'expectedResponseData' => [
+                    'error' => [
+                        'type' => 'invalid_request',
+                        'payload' => [
+                            'host-url' => [
+                                'value' => '',
+                                'message' => 'This value should not be blank.',
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ];
+    }
+}
