+++ tests/DataProvider/CreateSourceSuccessDataProviderTrait.php
@@ -1,0 +1,67 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\DataProvider;
+
+use App\Enum\Source\Type;
+use App\Request\FileSourceRequest;
+use App\Request\GitSourceRequest;
+use App\Request\SourceRequestInterface;
+
+trait CreateSourceSuccessDataProviderTrait
+{
+    /**
+     * @return array<mixed>
+     */
+    public function createSourceSuccessDataProvider(): array
+    {
+        $hostUrl = 'https://example.com/repository.git';
+        $path = '/';
+        $credentials = md5((string) rand());
+        $label = 'file source label';
+
+        return [
+            'git source, credentials missing' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
+                    GitSourceRequest::PARAMETER_HOST_URL => $hostUrl,
+                    GitSourceRequest::PARAMETER_PATH => $path
+                ],
+                'expected' => [
+                    'user_id' => TestConstants::AUTHENTICATED_USER_ID_PLACEHOLDER,
+                    'type' => Type::GIT->value,
+                    'host_url' => $hostUrl,
+                    'path' => $path,
+                    'has_credentials' => false,
+                ],
+            ],
+            'git source, credentials present' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
+                    GitSourceRequest::PARAMETER_HOST_URL => $hostUrl,
+                    GitSourceRequest::PARAMETER_PATH => $path,
+                    GitSourceRequest::PARAMETER_CREDENTIALS => $credentials,
+                ],
+                'expected' => [
+                    'user_id' => TestConstants::AUTHENTICATED_USER_ID_PLACEHOLDER,
+                    'type' => Type::GIT->value,
+                    'host_url' => $hostUrl,
+                    'path' => $path,
+                    'has_credentials' => true,
+                ],
+            ],
+            'file source' => [
+                'requestParameters' => [
+                    SourceRequestInterface::PARAMETER_TYPE => Type::FILE->value,
+                    FileSourceRequest::PARAMETER_LABEL => $label
+                ],
+                'expected' => [
+                    'user_id' => TestConstants::AUTHENTICATED_USER_ID_PLACEHOLDER,
+                    'type' => Type::FILE->value,
+                    'label' => $label,
+                ],
+            ],
+        ];
+    }
+}
