+++ tests/Functional/Controller/UserSourceControllerTest.php
@@ -8,25 +8,30 @@ use App\Entity\FileSource;
 use App\Entity\GitSource;
 use App\Entity\RunSource;
 use App\Entity\SourceInterface;
-use App\Enum\RunSource\FailureReason;
 use App\Enum\RunSource\State;
 use App\Enum\Source\Type;
 use App\Model\EntityId;
 use App\Repository\RunSourceRepository;
 use App\Repository\SourceRepository;
-use App\Request\FileSourceRequest;
-use App\Request\GitSourceRequest;
-use App\Request\SourceRequestInterface;
 use App\Services\RunSourceSerializer;
 use App\Services\Source\Store;
+use App\Tests\DataProvider\DeleteSourceSuccessDataProviderTrait;
+use App\Tests\DataProvider\GetSourceSuccessDataProviderTrait;
+use App\Tests\DataProvider\TestConstants;
+use App\Tests\DataProvider\UpdateSourceInvalidRequestDataProviderTrait;
+use App\Tests\DataProvider\UpdateSourceSuccessDataProviderTrait;
 use App\Tests\Model\UserId;
 use App\Tests\Services\EntityRemover;
 use App\Tests\Services\FileStoreFixtureCreator;
-use App\Tests\Services\SourceUserIdMutator;
 use League\Flysystem\FilesystemOperator;
 
 class UserSourceControllerTest extends AbstractSourceControllerTest
 {
+    use DeleteSourceSuccessDataProviderTrait;
+    use GetSourceSuccessDataProviderTrait;
+    use UpdateSourceInvalidRequestDataProviderTrait;
+    use UpdateSourceSuccessDataProviderTrait;
+
     private SourceRepository $sourceRepository;
     private RunSourceRepository $runSourceRepository;
     private Store $store;
@@ -99,7 +104,7 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
     }
 
     /**
-     * @dataProvider getSuccessDataProvider
+     * @dataProvider getSourceSuccessDataProvider
      *
      * @param array<mixed> $expectedResponseData
      */
@@ -114,71 +119,6 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
 
         $this->responseAsserter->assertSuccessfulJsonResponse($response, $expectedResponseData);
         $this->requestAsserter->assertAuthorizationRequestIsMade();
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function getSuccessDataProvider(): array
-    {
-        $userId = SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER;
-
-        $gitSource = new GitSource($userId, 'https://example.com/repository.git', '/', md5((string) rand()));
-        $fileSource = new FileSource($userId, 'file source label');
-        $runSource = new RunSource($fileSource);
-
-        $failureMessage = 'fatal: repository \'http://example.com/repository.git\' not found';
-        $failedRunSource = (new RunSource($gitSource))->setPreparationFailed(
-            FailureReason::GIT_CLONE,
-            $failureMessage
-        );
-
-        return [
-            Type::GIT->value => [
-                'source' => $gitSource,
-                'expectedResponseData' => [
-                    'id' => $gitSource->getId(),
-                    'user_id' => $gitSource->getUserId(),
-                    'type' => Type::GIT->value,
-                    'host_url' => $gitSource->getHostUrl(),
-                    'path' => $gitSource->getPath(),
-                    'has_credentials' => true,
-                ],
-            ],
-            Type::FILE->value => [
-                'source' => $fileSource,
-                'expectedResponseData' => [
-                    'id' => $fileSource->getId(),
-                    'user_id' => $fileSource->getUserId(),
-                    'type' => Type::FILE->value,
-                    'label' => $fileSource->getLabel(),
-                ],
-            ],
-            Type::RUN->value => [
-                'source' => $runSource,
-                'expectedResponseData' => [
-                    'id' => $runSource->getId(),
-                    'user_id' => $userId,
-                    'type' => Type::RUN->value,
-                    'parent' => $runSource->getParent()?->getId(),
-                    'parameters' => [],
-                    'state' => State::REQUESTED->value,
-                ],
-            ],
-            Type::RUN->value . ': preparation failed' => [
-                'source' => $failedRunSource,
-                'expectedResponseData' => [
-                    'id' => $failedRunSource->getId(),
-                    'user_id' => $userId,
-                    'type' => Type::RUN->value,
-                    'parent' => $failedRunSource->getParent()?->getId(),
-                    'parameters' => [],
-                    'state' => State::FAILED->value,
-                    'failure_reason' => FailureReason::GIT_CLONE->value,
-                    'failure_message' => $failureMessage,
-                ],
-            ],
-        ];
     }
 
     public function testUpdateUnauthorizedUser(): void
@@ -200,7 +140,7 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
     }
 
     /**
-     * @dataProvider updateInvalidRequestDataProvider
+     * @dataProvider updateSourceInvalidRequestDataProvider
      *
      * @param array<string, string> $payload
      * @param array<mixed>          $expectedResponseData
@@ -222,42 +162,7 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
     }
 
     /**
-     * @return array<mixed>
-     */
-    public function updateInvalidRequestDataProvider(): array
-    {
-        $userId = SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER;
-        $hostUrl = 'https://example.com/repository.git';
-        $path = '/';
-        $credentials = md5((string) rand());
-
-        $gitSource = new GitSource($userId, $hostUrl, $path, $credentials);
-
-        return [
-            Type::GIT->value . ' missing host url' => [
-                'source' => $gitSource,
-                'payload' => [
-                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
-                    GitSourceRequest::PARAMETER_HOST_URL => '',
-                    GitSourceRequest::PARAMETER_PATH => $path,
-                ],
-                'expectedResponseData' => [
-                    'error' => [
-                        'type' => 'invalid_request',
-                        'payload' => [
-                            'host-url' => [
-                                'value' => '',
-                                'message' => 'This value should not be blank.',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider updateDataProvider
+     * @dataProvider updateSourceSuccessDataProvider
      *
      * @param array<string, string> $payload
      * @param array<mixed>          $expectedResponseData
@@ -276,74 +181,6 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
 
         $this->responseAsserter->assertSuccessfulJsonResponse($response, $expectedResponseData);
         $this->requestAsserter->assertAuthorizationRequestIsMade();
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function updateDataProvider(): array
-    {
-        $userId = SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER;
-        $hostUrl = 'https://example.com/repository.git';
-        $path = '/';
-        $credentials = md5((string) rand());
-        $newHostUrl = 'https://new.example.com/repository.git';
-        $newPath = '/new';
-
-        $label = 'file source label';
-        $newLabel = 'new file source label';
-
-        $fileSource = new FileSource($userId, $label);
-        $gitSource = new GitSource($userId, $hostUrl, $path, $credentials);
-
-        return [
-            Type::FILE->value => [
-                'source' => $fileSource,
-                'payload' => [
-                    SourceRequestInterface::PARAMETER_TYPE => Type::FILE->value,
-                    FileSourceRequest::PARAMETER_LABEL => $newLabel,
-                ],
-                'expectedResponseData' => [
-                    'id' => $fileSource->getId(),
-                    'user_id' => $fileSource->getUserId(),
-                    'type' => Type::FILE->value,
-                    'label' => $newLabel,
-                ],
-            ],
-            Type::GIT->value . ' credentials present and empty' => [
-                'source' => $gitSource,
-                'payload' => [
-                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
-                    GitSourceRequest::PARAMETER_HOST_URL => $newHostUrl,
-                    GitSourceRequest::PARAMETER_PATH => $newPath,
-                    GitSourceRequest::PARAMETER_CREDENTIALS => null,
-                ],
-                'expectedResponseData' => [
-                    'id' => $gitSource->getId(),
-                    'user_id' => $gitSource->getUserId(),
-                    'type' => Type::GIT->value,
-                    'host_url' => $newHostUrl,
-                    'path' => $newPath,
-                    'has_credentials' => false,
-                ],
-            ],
-            Type::GIT->value . ' credentials not present' => [
-                'source' => $gitSource,
-                'payload' => [
-                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
-                    GitSourceRequest::PARAMETER_HOST_URL => $newHostUrl,
-                    GitSourceRequest::PARAMETER_PATH => $newPath,
-                ],
-                'expectedResponseData' => [
-                    'id' => $gitSource->getId(),
-                    'user_id' => $gitSource->getUserId(),
-                    'type' => Type::GIT->value,
-                    'host_url' => $newHostUrl,
-                    'path' => $newPath,
-                    'has_credentials' => false,
-                ],
-            ],
-        ];
     }
 
     public function testDeleteUnauthorizedUser(): void
@@ -365,7 +202,7 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
     }
 
     /**
-     * @dataProvider deleteSuccessDataProvider
+     * @dataProvider deleteSourceSuccessDataProvider
      */
     public function testDeleteSuccess(SourceInterface $source, int $expectedRepositoryCount): void
     {
@@ -379,32 +216,6 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
         $this->responseAsserter->assertSuccessfulResponseWithNoBody($response);
         $this->requestAsserter->assertAuthorizationRequestIsMade();
         self::assertSame($expectedRepositoryCount, $this->sourceRepository->count([]));
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function deleteSuccessDataProvider(): array
-    {
-        return [
-            Type::FILE->value => [
-                'source' => new FileSource(SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER, 'label'),
-                'expectedRepositoryCount' => 0,
-            ],
-            Type::GIT->value => [
-                'source' => new GitSource(
-                    SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER,
-                    'https://example.com/repository.git'
-                ),
-                'expectedRepositoryCount' => 0,
-            ],
-            Type::RUN->value => [
-                'source' => new RunSource(
-                    new FileSource(SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER, 'label')
-                ),
-                'expectedRepositoryCount' => 1,
-            ],
-        ];
     }
 
     public function testDeleteRunSourceDeletesRunSourceFiles(): void
@@ -515,7 +326,7 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
      */
     public function prepareSuccessDataProvider(): array
     {
-        $userId = SourceUserIdMutator::AUTHENTICATED_USER_ID_PLACEHOLDER;
+        $userId = TestConstants::AUTHENTICATED_USER_ID_PLACEHOLDER;
 
         $fileSource = new FileSource($userId, 'file source label');
         $gitSource = new GitSource($userId, 'https://example.com/repository.git', '/', md5((string) rand()));
