+++ tests/Integration/CreateSourceTest.php
@@ -5,16 +5,16 @@ declare(strict_types=1);
 namespace App\Tests\Integration;
 
 use App\Entity\SourceInterface;
-use App\Enum\Source\Type;
 use App\Repository\SourceRepository;
-use App\Request\FileSourceRequest;
-use App\Request\GitSourceRequest;
-use App\Request\InvalidSourceTypeRequest;
-use App\Request\SourceRequestInterface;
+use App\Tests\DataProvider\CreateSourceInvalidRequestDataProviderTrait;
+use App\Tests\DataProvider\CreateSourceSuccessDataProviderTrait;
 use App\Tests\Services\EntityRemover;
 
 class CreateSourceTest extends AbstractIntegrationTest
 {
+    use CreateSourceInvalidRequestDataProviderTrait;
+    use CreateSourceSuccessDataProviderTrait;
+
     private SourceRepository $sourceRepository;
 
     protected function setUp(): void
@@ -39,7 +39,7 @@ class CreateSourceTest extends AbstractIntegrationTest
     }
 
     /**
-     * @dataProvider createInvalidRequestDataProvider
+     * @dataProvider createSourceInvalidRequestDataProvider
      *
      * @param array<string, string> $requestParameters
      * @param array<string, string> $expectedResponseData
@@ -52,48 +52,7 @@ class CreateSourceTest extends AbstractIntegrationTest
     }
 
     /**
-     * @return array<mixed>
-     */
-    public function createInvalidRequestDataProvider(): array
-    {
-        return [
-            'invalid source type' => [
-                'requestParameters' => [
-                    SourceRequestInterface::PARAMETER_TYPE => 'invalid',
-                ],
-                'expectedResponseData' => [
-                    'error' => [
-                        'type' => 'invalid_request',
-                        'payload' => [
-                            'type' => [
-                                'value' => 'invalid',
-                                'message' => InvalidSourceTypeRequest::ERROR_MESSAGE,
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-            'git source missing host url' => [
-                'requestParameters' => [
-                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
-                ],
-                'expectedResponseData' => [
-                    'error' => [
-                        'type' => 'invalid_request',
-                        'payload' => [
-                            'host-url' => [
-                                'value' => '',
-                                'message' => 'This value should not be blank.',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider createSuccessDataProvider
+     * @dataProvider createSourceSuccessDataProvider
      *
      * @param array<string, string> $requestParameters
      * @param array<mixed>          $expected
@@ -113,59 +72,5 @@ class CreateSourceTest extends AbstractIntegrationTest
         $expected['user_id'] = $this->authenticationConfiguration->authenticatedUserId;
 
         $this->responseAsserter->assertSuccessfulJsonResponse($response, $expected);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function createSuccessDataProvider(): array
-    {
-        $hostUrl = 'https://example.com/repository.git';
-        $path = '/';
-        $credentials = md5((string) rand());
-        $label = 'file source label';
-
-        return [
-            'git source, credentials missing' => [
-                'requestParameters' => [
-                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
-                    GitSourceRequest::PARAMETER_HOST_URL => $hostUrl,
-                    GitSourceRequest::PARAMETER_PATH => $path
-                ],
-                'expected' => [
-                    'user_id' => self::AUTHENTICATED_USER_ID_PLACEHOLDER,
-                    'type' => Type::GIT->value,
-                    'host_url' => $hostUrl,
-                    'path' => $path,
-                    'has_credentials' => false,
-                ],
-            ],
-            'git source, credentials present' => [
-                'requestParameters' => [
-                    SourceRequestInterface::PARAMETER_TYPE => Type::GIT->value,
-                    GitSourceRequest::PARAMETER_HOST_URL => $hostUrl,
-                    GitSourceRequest::PARAMETER_PATH => $path,
-                    GitSourceRequest::PARAMETER_CREDENTIALS => $credentials,
-                ],
-                'expected' => [
-                    'user_id' => self::AUTHENTICATED_USER_ID_PLACEHOLDER,
-                    'type' => Type::GIT->value,
-                    'host_url' => $hostUrl,
-                    'path' => $path,
-                    'has_credentials' => true,
-                ],
-            ],
-            'file source' => [
-                'requestParameters' => [
-                    SourceRequestInterface::PARAMETER_TYPE => Type::FILE->value,
-                    FileSourceRequest::PARAMETER_LABEL => $label
-                ],
-                'expected' => [
-                    'user_id' => self::AUTHENTICATED_USER_ID_PLACEHOLDER,
-                    'type' => Type::FILE->value,
-                    'label' => $label,
-                ],
-            ],
-        ];
     }
 }
