+++ tests/Services/SourceUserIdMutator.php
@@ -6,12 +6,11 @@ namespace App\Tests\Services;
 
 use App\Entity\AbstractSource;
 use App\Entity\SourceInterface;
+use App\Tests\DataProvider\TestConstants;
 use webignition\ObjectReflector\ObjectReflector;
 
 class SourceUserIdMutator
 {
-    public const AUTHENTICATED_USER_ID_PLACEHOLDER = '{{ authenticated_user_id }}';
-
     public function __construct(
         private readonly AuthenticationConfiguration $authenticationConfiguration,
     ) {
@@ -19,7 +18,10 @@ class SourceUserIdMutator
 
     public function setSourceUserId(SourceInterface $source): SourceInterface
     {
-        if ($source instanceof AbstractSource && self::AUTHENTICATED_USER_ID_PLACEHOLDER == $source->getUserId()) {
+        if (
+            $source instanceof AbstractSource
+            && TestConstants::AUTHENTICATED_USER_ID_PLACEHOLDER == $source->getUserId()
+        ) {
             ObjectReflector::setProperty(
                 $source,
                 AbstractSource::class,
@@ -54,7 +56,7 @@ class SourceUserIdMutator
     {
         if (
             array_key_exists('user_id', $sourceData)
-            && self::AUTHENTICATED_USER_ID_PLACEHOLDER == $sourceData['user_id']
+            && TestConstants::AUTHENTICATED_USER_ID_PLACEHOLDER == $sourceData['user_id']
         ) {
             $sourceData['user_id'] = $this->authenticationConfiguration->authenticatedUserId;
         }
